backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
alembic==1.13.1
pydantic==2.7.4
pydantic-settings==2.3.4
redis==5.0.6
celery[redis]==5.4.0
httpx==0.27.0
python-jose[cryptography]==3.3.0
bcrypt==4.1.3
python-multipart==0.0.9
boto3==1.34.131
minio==7.2.7
pywebpush==2.0.0
py-vapid==1.9.2
